Added the AuxTel Visit integration test script.
Added the configuration for the AuxTel Visit test. Created the YAML formatting unit test for it. Imported the configs library in __init__.py.

Attempt at creating a ScriptQueue Controller.

Initial attempt at the AuxTel_Visit integration test unit test.

Added the auxtel_visit.py integration test script to setup.py.

Fixed the AuxTelVisit unit test to not have subprocess execute the script in the background, which was preventing it from communicating with the the controller and remote.

Added the ${PRODUCT_DIR}/bin folder to the PATH EnvVar in ups/ts_IntegrationTests.table.